+++ App.tsx
@@ -1,6 +1,7 @@
 import React from 'react';
 import { NavigationContainer } from '@react-navigation/native';
 import RootNavigator from './src/navigation/RootNavigator';
+import { StatusBar } from 'react-native';
    
 
 
@@ -8,6 +9,7 @@ export default function App() {
   return (
    <NavigationContainer>
      <RootNavigator/>
+     <StatusBar barStyle='dark-content'/>
    </NavigationContainer>
   );
 }
