+++ src/authentication/ForgetpasswordScreen.tsx
@@ -1,39 +1,49 @@
 import React from "react";
 import { View, Text, StyleSheet, Image } from "react-native";
-import CustomTextInput from "../components/TextInput/customTextInput";
-import CustomButton from "../components/button/customButton";
+import CustomTextInput from "../components/textInput/CustomTextInput";
+import CustomButton from "../components/button/CustomButton";
 import { Typography } from "../theme/Colors";
 import { assets } from "../../assets/images";
+import useAuthStore from "../stores/useAuthStore";
 import { useNavigation } from "@react-navigation/native";
 
-export default function VerifyotpScreen() {
+export default function ForgetpasswordScreen() {
   const Navigation = useNavigation();
 
   const Sendverification = () => {
-    Navigation.navigate("Passwordchange");
+    Navigation.navigate("Verifyotp");
   };
 
   return (
     <View style={styles.container}>
-      {/* <View style={{marginBottom:10}}> */}
-      <View style={styles.welcomeandsignup}>
-        <Text style={styles.welcomeText}>Enter Verification Code</Text>
-        <Text style={styles.subText}>
-          We will send you a message to set or reset your new
-        </Text>
-        <Text style={styles.subsubtext}>password</Text>
+      <View style={styles.logoContainer}>
+        <View style={styles.containerlogo}>
+          <View style={styles.diamond}>
+            <Image source={assets.logofirst} style={styles.diamond} />
+          </View>
+        </View>
       </View>
-      {/* </View> */}
+
+      <View style={styles.welcomeandsignup}>
+        <Text style={styles.welcomeText}>Forgot Password</Text>
+        <Text style={styles.subText}>
+          We will send you a message to set or reset your new{" "}
+        </Text>
+        <Text style={styles.subsubText}>password</Text>
+      </View>
 
       <CustomTextInput
         // value={email}
         // onChangeText={setEmail}
-        placeholder="Enter OTP here"
-        keyboardType="number-pad"
+        placeholder="Your Email / Phone Number"
+        keyboardType="email-address"
+        iconname="mail"
+        iconsize={25}
+        iconcolor={Typography.Colors.lightgrey}
       />
 
       <CustomButton
-        title="Confirm"
+        title="Send Verification"
         onPress={Sendverification}
         buttonStyle={styles.buttonstyle}
       />
@@ -43,8 +53,6 @@ export default function VerifyotpScreen() {
 
 const styles = StyleSheet.create({
   container: {
-    paddingTop: 200,
-    // marginBottom:60,
     flex: 1,
     paddingHorizontal: 20,
     backgroundColor: Typography.Colors.white,
@@ -53,7 +61,7 @@ const styles = StyleSheet.create({
     fontSize: 18,
     fontFamily: Typography.font.heavy,
     color: Typography.Colors.primary,
-    marginTop: 10,
+    marginTop: 20,
     marginBottom: 5,
   },
   subText: {
@@ -61,10 +69,11 @@ const styles = StyleSheet.create({
     fontSize: 15,
     color: Typography.Colors.lightgrey,
   },
-  subsubtext: {
+  subsubText: {
     fontFamily: Typography.font.regular,
     fontSize: 15,
     color: Typography.Colors.lightgrey,
+    // alignItems: "center",
     marginBottom: 30,
   },
   containerlogo: {
@@ -93,3 +102,20 @@ const styles = StyleSheet.create({
     height: 52,
   },
 });
+
+// useEffect(() => {
+//   const backHandler = BackHandler.addEventListener(
+//     'hardwareBackPress',
+//     () => {
+//       if (isSignUp === 'true') {
+//         navigation.navigate(SCREEN_URLS.signUp);
+//       } else {
+//         navigation.navigate(SCREEN_URLS.login);
+//       }
+
+//       return true;
+//     }
+//   );
+
+//   return () => backHandler.remove();
+// }, []);
